4.8.18 case study fix (#53)
* Ags cert (#45)

* Added debug logs

* added code to prevent cert. for ags casestudy

* Update CompositeSearchIndexerHelper.scala

---------


* Fix: Add null check for contentCache before logging Redis fetch details (#50)


* Optimized code for reading course data

* Updated logs and variable names to avoid confusion

---------
